requirements.txt
fastapi
itsdangerous
uvicorn[standard]
pydantic-settings
psycopg[binary]
Authlib
sqlalchemy
jinja2
python-multipart